Move tests for export components to separate files
